+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
-## [unreleased]
+## [1.0.0-alpha.9] - 2020-04-20
 
 ### Deprecated